use global memory pool

+++ apps/unit_tests/multi_cg/test_multi_cg_complex.cpp
@@ -49,7 +49,7 @@ struct BlockVector {
     }
 
     void repack(std::vector<int> const &ids) {
-        for (size_t i = 0; i < ids.size(); ++i) {
+        for (int i = 0; i < static_cast<int>(ids.size()); ++i) {
             auto j = ids[i];
             if (j != i) {
                 vec.col(i) = vec.col(j);
@@ -72,7 +72,7 @@ struct PosDefMatrixShifted {
     }
 
     void repack(std::vector<int> const &ids) {
-        for (size_t i = 0; i < ids.size(); ++i) {
+        for (int i = 0; i < static_cast<int>(ids.size()); ++i) {
             auto j = ids[i];
             if (j != i) {
                 shifts[i] = shifts[j];
